scripts/requirements.txt
# Requirements for scripts in this directory
PyYAML>=6.0

